Updated environment handling and UI components
Changed the title in `MainLayout.razor` to "Mimer Administration" and added a menu item for the current Mimer environment. Modified `Settings.razor` to replace the test data toggle with a conditional `MudIconButton` and introduced a `MudRadioGroup` for environment selection. Updated controls to disable based on the selected Mimer environment. Enhanced data loading in `Teams.razor` and `Users.razor` to reference `UserPreferences.MimerEnvironment`. Added a new `TestData` option in `MimerEnvironment.cs` and adjusted `CosmosService.cs` to fetch data based on the Mimer environment. Initialized `UserPreferences` to `TestData` by default and updated the interface accordingly.